feat: middleware validation
A Validation middleware was built to validate any field from the functions that will be used, it can receive any type and verify its field with the exactely path of the errror, making easy to the frontend side. Also, the configuration of the enviroment variables have been fixed at this commit.